requirements.txt
gpustat==1.1.1
torch==2.2.2
transformers==4.39.3
accelerate==0.29.3
tokenizers==0.15.0
sentencepiece==0.2.0
datasets==2.18.0
evaluate==0.4.1
fuzzywuzzy[speedup]==0.18.0
tree-sitter==0.21.3
tree-sitter-python==0.21.0
protobuf==5.26.1
openai==1.25.2
tiktoken==0.6.0
python-dotenv==1.0.1